Moved component to match file-structure of project
